requirements.txt
streamlit
google-cloud-bigquery
pandas